+++ src/app/modules/tables.php
@@ -12,12 +12,13 @@ class tablesModule extends zModule {
 
 	public function onEnabled() {
 		$this->z->core->includeCSS('resources/tables.css');
+		$this->z->core->includeCSS('resources/tables.css', false, 'admin.head');
 	}
 
 	public function createPaging() : zPaging {
 		return new zPaging(0, $this->getConfigValue('page_size'), $this->getConfigValue('max_pages_links'));
 	}
-	
+
 	public function createTable($entity_name = 'entity name', $view_name = null, $css = '') {
 		$table = new zTable($entity_name, $view_name, $css);
 		$table->paging = zPaging::getFromUrl(null, $this->getConfigValue('page_size'));
@@ -103,7 +104,7 @@ class tablesModule extends zModule {
 								foreach ($table->data as $row) {
 									$item_url = $this->z->core->url(sprintf($table->edit_link, $row->val($table->id_field_name)), $this->z->core->raw_path);
 									?>
-										<tr onclick="javascript:document.location = '<?=$item_url ?>';" class="">
+										<tr onclick="javascript:document.location = '<?=$item_url ?>';">
 											<?php
 												foreach ($table->fields as $field) {
 													?>