+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v5
 
 - `require_z_version` config option was dropped, now apps only specify `minimum_z_version`
+- admin now must specify it's own includes (names starting with `admin.`)